Refactor refrigerator and freezer to inheritance.
