+++ lib/chiller.rb
@@ -1,34 +1,8 @@
-class Chiller
+require_relative 'refrigerator_block'
 
-  ROOM_TEMPERATURE = 70
-
-  attr_reader :capacity, :temperature
-
-  def initialize(capacity = 100)
-    @capacity = capacity
-    @temperature = ROOM_TEMPERATURE
-    @power = :off
-    @contents = []
-  end
-
-  def turn_on
-    @power = :on
-  end
-
-  def turn_off
-    @power = :off
-  end
-
-  def add(item)
-    @contents << item
-  end
-
-  def remaining_capacity
-    capacity - @contents.map(&:volume).reduce(:+).to_i
-  end
+class Freezer < RefrigeratorBlock
 
   def set_level(level)
-    @temperature = ROOM_TEMPERATURE - level * 5
+    super(level, 5)
   end
-
 end
